+++ .replit
@@ -1,2 +1,2 @@
-language = "cpp"
-run="export CC=clang; export CXX=clang++; mkdir -p ../build; cd ../build; cmake ../acc-cosc-1337-spring-2021-Hannah-Kim-acc/; make; cd src/classwork/04_assign; ./cw_04_exe -s" 
+10language = "cpp"
+run="export CC=clang; export CXX=clang++; mkdir -p ../build; cd ../build; cmake ../acc-cosc-1337-spring-2021-Hannah-Kim-acc/; make; cd test/homework_test/01_variables_test; ./hw_01_tests -s"
