+++ src/classwork/02_assign/expressions.h
@@ -1,6 +1,2 @@
 //headerd
 
-double get_sales_tax_amount(double meal_amount);
-double get_tip_amount(double meal_amount,double tip_rate);
-
-
