+++ src/classwork/CMakeLists.txt
@@ -1,6 +1,6 @@
-ADD_SUBDIRECTORY(01_assign)
-ADD_SUBDIRECTORY(02_assign)
-ADD_SUBDIRECTORY(03_assign)
+#ADD_SUBDIRECTORY(01_assign)
+#ADD_SUBDIRECTORY(02_assign)
+#ADD_SUBDIRECTORY(03_assign)
 ADD_SUBDIRECTORY(04_assign)
 #ADD_SUBDIRECTORY(05_assign)
 #ADD_SUBDIRECTORY(06_assign)
