+++ src/homework/01_variables/variables.cpp
@@ -10,5 +10,16 @@ int add_numbers(int num1, int num2)
 //function returns product of num1 and num2
 
 
+double get_sales_tax_amount(double meal_amount)
+{
+  const double tax_rate = .0675;
+  return meal_amount * tax_rate;
+}
+double get_tip_amount(double meal_amount,double tip_rate)
+{
+  return meal_amount * tip_rate;
+}
+
+
