+++ src/homework/01_variables/variables.h
@@ -3,6 +3,10 @@ int add_numbers(int num1, int num2);
 
 //write function prototype for multiply_numbers with two integer parameters num1, and num2
 
+double get_sales_tax_amount(double meal_amount);
+double get_tip_amount(double meal_amount,double tip_rate);
+
+
