+++ test/homework_test/01_variables_test/01_variables_tests.cpp
@@ -12,3 +12,12 @@ TEST_CASE("Verify sum_numbers function")
 	REQUIRE(add_numbers(10, 10) == 20);
 }
 
+TEST_CASE("Test Get Sales Tax", "verification") {
+	REQUIRE(get_sales_tax_amount(10) == .675);
+  REQUIRE(get_sales_tax_amount(20) == 1.35);
+}
+
+TEST_CASE("Test Get Tip Amount", "verification") {
+	REQUIRE(get_tip_amount(20,0.15) == 3 );
+  REQUIRE(get_tip_amount(20,0.20) == 4);
+}
